--- Tests/Listener/EntityNotFoundException.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace JavierEguiluz\Bundle\EasyAdminBundle\Tests\Listener;
-
-/**
- * Test class for event listener test
- */
-class EntityNotFoundException extends \Exception
-{
-}
